Rename SiteHydroShareArchivalModal to HydroShareFormCard
